Move `stringutil`.
