+++ util/stringutil/scrub.go
