+++ variables/builtin.go
@@ -9,7 +9,7 @@ import (
 	"github.com/earthly/earthly/gitutil"
 	"github.com/earthly/earthly/llbutil"
 	"github.com/earthly/earthly/states/dedup"
-	"github.com/earthly/earthly/stringutil"
+	"github.com/earthly/earthly/util/stringutil"
 	specs "github.com/opencontainers/image-spec/specs-go/v1"
 )
